+++ src/Garn.hs
@@ -67,6 +67,7 @@ runWith env (WithGarnTsOpts garnConfig opts) = do
           forM_ packages $ \package -> do
             c <- runNix (NoStdin, "build", ".#" <> package)
             when (c /= ExitSuccess) $ exitWith c
+        TargetConfigEnvironment _ -> pure ()
         TargetConfigPackage (PackageTarget {}) -> do
           c <- runNix (NoStdin, "build", ".#" <> asNixFacing target)
           when (c /= ExitSuccess) $ exitWith c
@@ -87,6 +88,7 @@ checkTarget targetConfig = case targetConfig of
     forM_ checks $ \check -> do
       c <- runNix (NoStdin, "build", ".#checks." <> system <> "." <> check)
       when (c /= ExitSuccess) $ exitWith c
+  TargetConfigEnvironment _ -> pure ()
   TargetConfigPackage _ -> pure ()
   TargetConfigExecutable _ -> pure ()