+++ src/Garn/GarnConfig.hs
@@ -76,6 +76,7 @@ type Targets = Map TargetName TargetConfig
 
 data TargetConfig
   = TargetConfigProject ProjectTarget
+  | TargetConfigEnvironment EnvironmentTarget
   | TargetConfigPackage PackageTarget
   | TargetConfigExecutable ExecutableTarget
   deriving (Generic, Eq, Show)
@@ -85,6 +86,11 @@ data ProjectTarget = ProjectTarget
     packages :: [String],
     checks :: [String],
     runnable :: Bool
+  }
+  deriving (Generic, FromJSON, Eq, Show)
+
+data EnvironmentTarget = EnvironmentTarget
+  { description :: Maybe String
   }
   deriving (Generic, FromJSON, Eq, Show)
 
@@ -103,15 +109,17 @@ instance FromJSON TargetConfig where
     tag <- o .: fromString "tag"
     case tag of
       "project" -> TargetConfigProject <$> genericParseJSON defaultOptions (Object o)
+      "environment" -> TargetConfigEnvironment <$> genericParseJSON defaultOptions (Object o)
       "package" -> TargetConfigPackage <$> genericParseJSON defaultOptions (Object o)
       "executable" -> TargetConfigExecutable <$> genericParseJSON defaultOptions (Object o)
       _ -> fail $ "Unknown target tag: " <> tag
 
-getDescription :: TargetConfig -> String
+getDescription :: TargetConfig -> Maybe String
 getDescription = \case
-  TargetConfigProject (ProjectTarget {description}) -> description
-  TargetConfigPackage (PackageTarget {description}) -> description
-  TargetConfigExecutable (ExecutableTarget {description}) -> description
+  TargetConfigProject (ProjectTarget {description}) -> Just description
+  TargetConfigEnvironment (EnvironmentTarget {description}) -> description
+  TargetConfigPackage (PackageTarget {description}) -> Just description
+  TargetConfigExecutable (ExecutableTarget {description}) -> Just description
 
 newtype ReadConfigError = ReadConfigError {errorMessage :: String}
   deriving stock (Eq, Show)