+++ src/Garn/Optparse.hs
@@ -14,6 +14,7 @@ where
 
 import Control.Exception (throw)
 import qualified Data.Map as Map
+import Data.Maybe (fromMaybe)
 import Garn.Common (garnCliVersion)
 import qualified Garn.Errors
 import Garn.GarnConfig
@@ -107,6 +108,7 @@ buildCommand targets =
   where
     isBuildable = \case
       TargetConfigProject _ -> True
+      TargetConfigEnvironment _ -> False
       TargetConfigPackage _ -> True
       TargetConfigExecutable _ -> False
 
@@ -120,12 +122,19 @@ runCommand targets =
     isRunnable :: TargetConfig -> Bool
     isRunnable = \case
       TargetConfigProject projectTarget -> runnable projectTarget
+      TargetConfigEnvironment _ -> False
       TargetConfigPackage _ -> False
       TargetConfigExecutable _ -> True
 
 enterCommand :: Targets -> Parser WithGarnTsCommand
 enterCommand targets =
-  Enter <$> commandOptionsParser (Map.filter isProject targets)
+  Enter <$> commandOptionsParser (Map.filter isEnterable targets)
+  where
+    isEnterable = \case
+      TargetConfigProject _ -> True
+      TargetConfigEnvironment _ -> True
+      TargetConfigPackage _ -> False
+      TargetConfigExecutable _ -> False
 
 checkCommand :: Targets -> Parser WithGarnTsCommand
 checkCommand targets =
@@ -138,6 +147,7 @@ checkCommand targets =
 isProject :: TargetConfig -> Bool
 isProject = \case
   TargetConfigProject _ -> True
+  TargetConfigEnvironment _ -> False
   TargetConfigPackage _ -> False
   TargetConfigExecutable _ -> False
 
@@ -186,7 +196,7 @@ commandOptionsParser targets =
               (asUserFacing target)
               ( info
                   (pure (CommandOptions target targetConfig))
-                  (progDesc $ getDescription targetConfig)
+                  (maybe mempty progDesc (getDescription targetConfig))
               )
         )
         $ Map.assocs targets