+++ ts/javascript/mod.ts
@@ -57,7 +57,6 @@ export function mkNpmProject(args: {
   src: string;
   nodeVersion: NodeVersion;
 }): Project & {
-  devShell: Environment;
   node_modules: Package;
 } {
   const { pkgs, nodejs } = fromNodeVersion(args.nodeVersion);
@@ -95,7 +94,6 @@ export function mkNpmProject(args: {
       defaultEnvironment: devShell,
     },
     {
-      devShell,
       node_modules,
     },
   );
@@ -174,7 +172,6 @@ export function mkYarnProject(args: {
     },
     {
       pkg,
-      devShell,
       startDev,
     },
   );