+++ CHANGELOG.md
@@ -8,15 +8,15 @@
 - Make `executables` a list, and optional, in `mkHaskellPackage`
 - Add `addCabalExecutable` to Haskell projects
 - Make `garn init` on Haskell projects fill in the `executables` based on the cabal file.
-
 - For git repositories, add untracked files to build sandboxes for `Check`s and
   `Package`s.
-
 - Tweak the API for creating `Environment`s, making it easier to create
   `Environment`s that have a source directory
-
 - Fix needing to use `--` twice in `garn run` in order to pass arguments to
   executables.
+- Allow entering top-level `Environment`s with `garn enter`.
+- `Environment.mkNpmProject` and `Environment.mkYarnProject` don't expose a
+  `devShell` field anymore, since that was a duplicate of `defaultEnvironment`.
 
 - Create a new `edit` command. This opens VSCodium with deno LSP already set up,
   in exactly the same way as the old `editGarnConfig` function did. (That